Split authentication into separate middleware files
