+++ projects/user-telemetry-client/src/authentication/AuthCookie.ts
@@ -1,0 +1,6 @@
+export function cookieAuthentication(requestInit: RequestInit): RequestInit {
+        return {
+            ...requestInit,
+            credentials: "include"
+        }
+}
