+++ projects/user-telemetry-client/src/index.ts
@@ -1,2 +1,4 @@
 export { IUserTelemetryEvent } from "../../definitions/IUserTelemetryEvent"
-export { UserTelemetryEventSender } from "./TelemetryService"
+export { UserTelemetryEventSender, GuAuthMiddleware } from "./TelemetryService"
+export  { cookieAuthentication } from "./authentication/AuthCookie"
+export  { hmacAuthentication } from "./authentication/HmacRequest"
